# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type IngredientCategory = typeof IngredientCategory[keyof typeof IngredientCategory];


export const IngredientCategory = {
  produce: 'produce',
  dairy: 'dairy',
  meat_seafood: 'meat_seafood',
  bakery: 'bakery',
  pantry: 'pantry',
  frozen: 'frozen',
  beverages: 'beverages',
  spices: 'spices',
  other: 'other',
} as const;

export type MealSlot = typeof MealSlot[keyof typeof MealSlot];


export const MealSlot = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner',
  snack: 'snack',
} as const;

export interface Ingredient {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  unit?: string | null;
  category: IngredientCategory;
}

export interface Recipe {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  ingredients: Ingredient[];
  instructions: string;
  /** @nullable */
  servings: number | null;
  /** @nullable */
  prepMinutes: number | null;
  /** @nullable */
  cookMinutes: number | null;
  tags: string[];
  /** @nullable */
  photoUrl: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface RecipeInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  ingredients: Ingredient[];
  /** @minLength 1 */
  instructions: string;
  /** @minimum 1 */
  servings?: number;
  /** @minimum 0 */
  prepMinutes?: number;
  /** @minimum 0 */
  cookMinutes?: number;
  tags?: string[];
  photoUrl?: string;
}

export interface RecipeUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  ingredients?: Ingredient[];
  /** @minLength 1 */
  instructions?: string;
  /**
     * @minimum 1
     * @nullable
     */
  servings?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  prepMinutes?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  cookMinutes?: number | null;
  tags?: string[];
  /** @nullable */
  photoUrl?: string | null;
}

export interface RecipeSummary {
  id: number;
  title: string;
  /** @nullable */
  photoUrl: string | null;
  /** @nullable */
  servings: number | null;
  /** @nullable */
  prepMinutes: number | null;
  /** @nullable */
  cookMinutes: number | null;
  tags: string[];
}

export interface MealPlanEntry {
  id: number;
  date: string;
  mealSlot: MealSlot;
  /** @nullable */
  servings: number | null;
  recipe: RecipeSummary;
  createdAt: string;
}

export interface MealPlanEntryInput {
  date: string;
  mealSlot: MealSlot;
  recipeId: number;
  /** @minimum 1 */
  servings?: number;
}

export interface MealPlanEntryUpdate {
  date?: string;
  mealSlot?: MealSlot;
  recipeId?: number;
  /**
     * @minimum 1
     * @nullable
     */
  servings?: number | null;
}

export type GroceryListItemSource = typeof GroceryListItemSource[keyof typeof GroceryListItemSource];


export const GroceryListItemSource = {
  auto: 'auto',
  manual: 'manual',
} as const;

export interface GroceryListItem {
  id: number;
  weekStart: string;
  name: string;
  /** @nullable */
  quantity: number | null;
  /** @nullable */
  unit: string | null;
  category: IngredientCategory;
  checked: boolean;
  source: GroceryListItemSource;
  createdAt: string;
}

export interface GroceryListItemInput {
  weekStart: string;
  /** @minLength 1 */
  name: string;
  quantity?: number;
  unit?: string;
  category: IngredientCategory;
}

export interface GroceryListItemUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  unit?: string | null;
  category?: IngredientCategory;
  checked?: boolean;
}

export interface GenerateGroceryListInput {
  weekStart: string;
}

export interface DashboardSummary {
  weekStart: string;
  recipeCount: number;
  mealsPlannedThisWeek: number;
  groceryItemCount: number;
  groceryCheckedCount: number;
  recentRecipes: RecipeSummary[];
  todayEntries: MealPlanEntry[];
}

export type ListRecipesParams = {
/**
 * Case-insensitive match against title, tags, and ingredient names
 */
search?: string;
/**
 * Filter to recipes containing this exact tag
 */
tag?: string;
};

export type ListMealPlanEntriesParams = {
startDate: string;
endDate: string;
};

export type ListGroceryListItemsParams = {
/**
 * Monday date (YYYY-MM-DD) identifying the week
 */
weekStart: string;
};

export type GetDashboardSummaryParams = {
weekStart: string;
};