# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealPlanEntry } from './mealPlanEntry';
import type { RecipeSummary } from './recipeSummary';

export interface DashboardSummary {
  weekStart: Date;
  recipeCount: number;
  mealsPlannedThisWeek: number;
  groceryItemCount: number;
  groceryCheckedCount: number;
  recentRecipes: RecipeSummary[];
  todayEntries: MealPlanEntry[];
}