# lib/api-zod/src/generated/types/groceryListItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroceryListItemSource } from './groceryListItemSource';
import type { IngredientCategory } from './ingredientCategory';

export interface GroceryListItem {
  id: number;
  weekStart: Date;
  name: string;
  /** @nullable */
  quantity: number | null;
  /** @nullable */
  unit: string | null;
  category: IngredientCategory;
  checked: boolean;
  source: GroceryListItemSource;
  createdAt: Date;
}